Add system prop to disable templ caching for #417
TemplateEngine implementations perform caching based on whether or not a
system property is defined (see CachingTemplateEngine for property
name). Also, TemplateEngine implementations have isCachingEnabled() to
provide a read-only API view of this property for the TemplateEngine.
The `cache` field in CachingTemplateEngine is still initialized even if
caching is disable. This was kept to minimized future support for adding
a setCachingEnabled(boolean) to TemplateEngines to enable callers to
change the caching property during runtime and to avoid having to
re-initialized the `cache` field.