Support named path segments and regular expressions with the SockJSHandler Adds router implementation PrefixedRouterImpl which automatically prefixes routes added to it with a static prefix. An instance can be passed into SockJSHandler.create() which is then mounted on the parent of the prefix.
See PrefixedRouterImpl for an example.

Previously it wasn't possible to use named path segment or regular expressions
in the path where a SockJSHandler is mounted. Only literal prefix values
were supported, because SockJSHandler always created its own router
and used this to handle requests.

(It wasn't possible to implement just the router interface in PrefixedRouterImpl
as its necessary to override the handle* methods which need implementations like
RouteImpl which are private in RouterImpl.)