Changed exception message (#1274)
The form login handler checks if the request is set to expect multipart.
If this property of the request is set to false, an exception is thrown.
Previously, the exception's description message suggested adding a
BodyHandler. The new message suggests setting expect multipart to true.

(cherry picked from commit 384a2ced197599437990348b77f8d14bdbbb484f)