add a Vertx wrapper for the MappedBatchLoaderWithContext
This follows the pattern started by wrapping the BatchLoaderWithContext in a
Vertx callback friendly API. Instead of receiving a list of keys and returning
a list of values, this will accept a set of keys and return a map of key-value
pairs. This can be useful for maintaining integrity when a 1-1 mapping can't
be guaranteed, or when parial loads are expected.

https://github.com/graphql-java/java-dataloader#returning-a-map-of-results-from-your-batch-loader