backport of #1541 - Add a Lazy Session option for the SessionHandler
Backport of #1541.

The goal of this commit is to have a mean to stop vert.x SessionHandler to always create a session Cookie.
In some cases, application built with Vert.x doesn't know in advance if they will use the session, and if not, may not want to be treated like a stateful application by network components (ie Load Balancers).
This commit follows the discussion accessible here : https://groups.google.com/d/msg/vertx/hoQEgU2d_FY/uUu2jdd7AQAJ

This commit adds a new SessionHandler option : lazySession. With this option set, the session is still (re)created in the SessionHandler.handle(...) but not always stored in the sessionStore.
The RoutingContext is now responsible to track whether the session has been accessed. In that case the new accessedSession() method must return true.
Thanks to this new RoutingContext.accessedSession() method, the SessionHandler is able to know whether the session has been accessed and to decide whether it has to store it.

Signed-off-by: Nils Renaud <renaud.nils@gmail.com>