Initialize fully HttpContext's request options during the PREPARE_REQUEST phase instead of the CREATE_REQUEST phase.
Previously the initialization of the HttpCOntext request options was started in the PREPARE_REQUEST and finished in the CREATE_REQUEST. The code is entirely moved to the PREPARE_REQUEST so an interceptor can rely on this behavior and update the request options between the two phases. Now the HttpContext assumes that the PREPARE_REQUEST phase creates the request options and the CREATE_REQUEST creates the HttpClientRequest from the request options.

Consequently HTTP redirect is adapted to initalize fully the request options before branching to the CREATE_REQUEST phase. Previously HTTP redirect was simply reusing the request options provided from the redirect handler and relying on the fact that the CREATE_REQUEST was finishing the request option initialization.

Signed-off-by: Julien Viet <julien@julienviet.com>