Reset interceptor index always after call HttpContext `fire` function
Now the interceptor index is restarted only when any of the methods using
`fire(ClientPhase phase)` are called from within the interceptor `handle(X)`,
but if `fire(ClientPhase phase)` is executed from outside the interceptor
`handle(X)` then the interceptors before the last executed interceptor and
this interceptor itself are ignored for the new `phase`.

This change modifies the current logic to always reset the interceptor
index.

Signed-off-by: Juan Manuel Fernández García-Minguillán <juanmanuel.fdez.gm@gmail.com>