Add an "advanced" LoggerFormatter API to expose more request metadata
As LoggerHandlerImpl already calculated this important information, that
is otherwise hard/impossible to get (e.g. request start time), so if we
want to use that metadata in a formatter (e.g. to mimic
LoggerFormat.DEFAULT) - it'd be best if we don't need to repeat work
that had already been done.