Check for computed content length before using chunked encoding
Currently, the code only checks if content length has been set in the initial request, and ignores any content length computed from the request body. This means that when using form data (sendForm()), even though a content length is computed and added to requestOptions, the request itself will always be sent in chunked mode unless content-length was added to the request earlier (even with an incorrect value).