Document MVEL limitation regarding inclusions
You may include a template file into an MVEL template using `@include{}`, as long as the templates reside on the filesystem.

Template inclusion doesn't work if the templates are loaded from the classpath using the Vert.x file resolver.

This is a limitation of the MVEL engine: some other engines let Vert.x plug in a custom template file resolver, MVEL doesn't.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>