Support MVEL variable resolver again (#2548)
There was a regression introduced in #468. The user-supplied variables were passed to the template engine with a MapVariableResolverFactory and we switched to a context object.

For this reason, it was no longer possible to declare variables in template files.

In this PR, we invoke the template engine with a MapVariableResolverFactory again, instead of the ImmutableDefaultFactory (which prevents from declaring variables).

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>