CORS: support Chrome extensions
See #2556

This commit makes it possible to use star origin in CORS handler setup with a Chrome extension client app.
Chrome Extensions use an origin of the follow form: chrome-extension://<extension-id>
An extension id is 32 chars long and contains chars from 'a' to 'p' (0->9 + A->F translated to a->p).

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>