Intermittent test failures in Healthcheck tests and StaticHandlerTest
Fix #2551

In StaticHandlerTest, some tests create alternative http servers. We must make sure the original and the replacement servers are closed properly.
Otherwise, the test results may be affected by the test execution order.

In HealthCheckTestBase, we make sure the client, the server and the Vert.x server are closed properly, and in this order.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>