Fix for SessionHandlerTestBase#doTestSessionRetryTimeout
Since session flushing is an asynchronous process, the session data may not be removed from the store when the next request is handled.

With manual flushing, we make sure the session store has destroyed the data before responding to the request.
So the next one never gets stale session data.

This will help with Infinispan SharedData map session store testing.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>