Improve ProtocolUpgradeHandler priority checking
See #2676

ProtocolUpgradeHandler (GraphQLWS, Proxy) instances must not be installed after the BodyHandler.
However, they are often used with authentication handlers (e.g. Basic auth handler), because authentication can be required before allowing the connection upgrade.
The problem is that, currently, authn/authz handlers have higher priorities than body handlers (for good reasons, we may need to inspect the body to perform form login).
So, if a user tries to install a ProtocolUpgradeHandler after an auth handler, an ISE is thrown.

With this change, the order of priorities is not changed.
Instead, there's a special check for ProtocolUpgradeHandler instances: the BodyHandler must not be found in the already installed handlers.

Also, the warning about ProtocolUpgradeHandler being installed before another handler has been removed.
Indeed, it is a fairly common case to hand over the request to another handler if the protocol upgrade has not been made (e.g. GraphQL handler installed after the GraphQLWS handler).

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>