Replace free cast in TemplateEngine#unwrap with covariant return type.
Motivation:

The TemplateEngine#unwrap uses a free downcast pattern (<T> T) that does not indicate the actual engine type and force the developer to know about the actual type to cast to.

This can be replaced by covariant return type that carries the exact type of the unwrapped engine.

Changes:

- remove generic <T> declaration in unwrap
- template engine supporting it, overrides this method with the actual engine type