Upgrade JTE template engine to 3.2.1
Maven toolchain is used to compile with JDK17 (required) when Maven is executed with JDK11

JTE Maven plugin doesn't support toolchains so it's executed when running on JDK17+.
As a consequence tests can be run only when running on JDK17+.

On GitHub Actions, JTE tests will be executed separately.
Enabled by default, they are disabled in ci-5.x.yml when testing with JDK11.

Note that deployment in ci-5.x.yml is executed with JDK17 because Maven toolchains are not easy to set up on GitHub Actions.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>